Annotate max_delay as int instead of float

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -10,13 +10,13 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: float = 10) -> float:
+async def wait_random(max_delay: int = 10) -> float:
     """
     Asynchronous coroutine that waits for a random delay between 0 and
     max_delay (inclusive) seconds.
 
     Args:
-        max_delay (float, optional): The maximum delay in seconds.
+        max_delay (int, optional): The maximum delay in seconds.
         Defaults to 10.
 
     Returns:
